Make the JUnit5 runner compatible with Java 8
This requires getting rid of `var` and polyfilling a few unavailable
methods. Since compiling with a fixed `--release` is required to prevent
breakages when `SecurityManager` goes away in future releases, this also
includes a workaround for https://bugs.openjdk.org/browse/JDK-8214165.

In order to pay back for the additional tech debt, this commit simplies
the build of the runner by using `exports` and `neverlink`.